+++ src/ModelingData/TKBRep/BRepGraph/BRepGraph_History.hxx
@@ -74,12 +74,13 @@ public:
 
   //! Number of recorded history events.
   //! @return record count
-  [[nodiscard]] Standard_EXPORT int NbRecords() const;
+  [[nodiscard]] Standard_EXPORT size_t NbRecords() const;
 
   //! Access a record by index (0-based).
   //! @param[in] theRecordIdx zero-based index into the records vector
   //! @return the history record at the given index
-  [[nodiscard]] Standard_EXPORT const BRepGraph_HistoryRecord& Record(const int theRecordIdx) const;
+  [[nodiscard]] Standard_EXPORT const BRepGraph_HistoryRecord& Record(
+    const size_t theRecordIdx) const;
 
   //! Enable or disable history recording.
   //! @param[in] theVal true to enable, false to disable
